Move FileUtils to support package.
